% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_thredds_url.R
\name{build_thredds_url}
\alias{build_thredds_url}
\title{Build an object containing URL and date information.}
\usage{
build_thredds_url(
  df,
  datetime = "datecollected",
  full_timeperiod = FALSE,
  url = "",
  path = "",
  file = "",
  var = "",
  verbose = TRUE
)
}
\arguments{
\item{df}{The dataframe from which date data will be derived.}

\item{datetime}{The column in df that contains date/time information.}

\item{url}{The base URL of the appropriate THREDDS server.}

\item{path}{The path to the file on the THREDDS server, not including the file name. (This lets us expand the functionality to
iterate over multiple folders).}

\item{file}{The name of the file on the THREDDS server.
Might need to expand this to include dates if the need arises.}
}
\value{
A URL (or tibble thereof) indicating files to be pulled.
}
\description{
This function creates an object containing relevate date and URL information for
pulling environmental variables from an arbitrary server.
}
\details{
Built to be a small-scale utility function to grab an arbitrary file from an abritrary THREDDS server. May expand to allow
iteration over multiple files.
}
\keyword{internal}
