% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{choose_file}
\alias{choose_file}
\title{Choose a file interactively}
\usage{
choose_file(caption)
}
\arguments{
\item{caption}{caption to display in choose file window. If method defaults to base::file.choose then
caption is printed to the console (doesn't work reliably due to shiny server issues)}
}
\description{
Provides a platform-independent way to interactively select files
}
