% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAndFetchPolygons.R
\name{createAndFetchPolygons}
\alias{createAndFetchPolygons}
\title{Fetch species home range polygons if they exist locally, create and store them if not.}
\usage{
createAndFetchPolygons(
  sciNames,
  occurrenceFolder = "localOccurrenceData/",
  polygonFolder = "localPolygonData/"
)
}
\description{
Fetch species home range polygons if they exist locally, create and store them if not.
}
