% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputList.R
\name{createInputList}
\alias{createInputList}
\title{Create an iterable list of file paths that can be looped over in the Remora notebook.}
\usage{
createInputList(path)
}
\arguments{
\item{path}{The path to either the file or the folder we're going to operate on.}
}
\value{
Returns either a character vector of filenames or NULL if 'path' is improperly formatted.
}
\description{
Since the Remora R notebook needs to be able to process a whole folder worth of files rather than a single file, this helper function will produce a list over which the notebook can
iterate. This lets us change as little of the underlying structure as possible while still allowing us to do batch processing.
}
