% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDownload.R
\name{dataDownload}
\alias{dataDownload}
\title{Download Bluelink environmental data}
\usage{
dataDownload(type, year, month = NULL, dir, varname, quiet = TRUE)
}
\arguments{
\item{type}{type of environmental data to download, "day" or "month" resolutions currently accepted.}

\item{year}{year of interest to download data for. Must be provided}

\item{month}{month of interest to download data for. If not provided, all months with available data will be downloaded.}

\item{varname}{Bluelink variable of interest to be downloaded.}

\item{quiet}{Should the files be downloaded in silence?}
}
\value{
Download netCDF files from Bluelink for internal processing.
}
\description{
helper function to download netCDF files, not called by user
}
\keyword{internal}
