% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn_imos_column_map.R
\name{derive_rcvr_from_det}
\alias{derive_rcvr_from_det}
\title{Derive a receiver metadata data.frame from OTN-formatted data to IMOS-format}
\usage{
derive_rcvr_from_det(det_dataframe)
}
\arguments{
\item{det_dataframe}{The dataframe containing detection information.
Most likely a detection extract.}
}
\description{
Hack together a piecemeal receiver metadata dataframe for instances
where we get detection data, no receiver/tag metadata, but still want to act
as though we DID get receiver/tag metadata.
}
\keyword{internal}
