% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{det_per_receiver_location}
\alias{det_per_receiver_location}
\title{Receiver location summary}
\usage{
det_per_receiver_location(data)
}
\arguments{
\item{data}{merged detections data and receiver metadata}
}
\value{
a tibble
}
\description{
The data needs to have the detections merged to the receiver metadata.
Not to be called by user
}
