% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{det_per_species}
\alias{det_per_species}
\title{Detections per species}
\usage{
det_per_species(data)
}
\arguments{
\item{data}{detections data}
}
\value{
a tibble
}
\description{
not to be called by user
}
