% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{det_per_tag}
\alias{det_per_tag}
\title{Number of detections per transmitter}
\usage{
det_per_tag(data)
}
\arguments{
\item{data}{detections data}
}
\value{
a tibble
}
\description{
not to be called by user
}
