% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{det_per_tag_location}
\alias{det_per_tag_location}
\title{Number of detections per tag, tag deployment coordinates and datetime}
\usage{
det_per_tag_location(data)
}
\arguments{
\item{data}{detections data}
}
\value{
a tibble
}
\description{
not to be called by user
}
