% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{det_species_station}
\alias{det_species_station}
\title{Detections per station per species}
\usage{
det_species_station(data)
}
\arguments{
\item{data}{merged detections data and receiver metadata}
}
\value{
a tibble
}
\description{
not to be called by user
}
