% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_urls.R, R/build_urls_arbitrary.R
\name{.build_urls}
\alias{.build_urls}
\title{Build urls to pull data environmental data from IMOS THREDDS server}
\usage{
.build_urls(dates, var_name, verbose = TRUE)

.build_urls(dates, var_name, verbose = TRUE)
}
\arguments{
\item{dates}{detection data source in data frame with at the minimum a X, Y and date time field}

\item{var_name}{variable needed options include available options ('rs_sst', 'rs_sst_interpolated',
'rs_salinity', 'rs_chl', 'rs_turbidity', 'rs_npp', 'rs_current'). For 'bathy' and 'dist_to_land', .pull_var() downloads
them directly without the need for this subroutine.}

\item{verbose}{should function provide details of what operation is being conducted. Set to \code{FALSE} to keep it quiet}
}
\value{
a tibble with urls from where to download environmental data and layer names (where applicable)

a tibble with urls from where to download environmental data and layer names (where applicable)
}
\description{
Internal function to check, define and create urls for the .pull_var() subroutine

Internal function to check, define and create urls for the .pull_var() subroutine
}
\details{
Internal function to build URLs for specific environmental variables and dates from IMOS Thredds server (http://thredds.aodn.org.au/)

Internal function to build URLs for specific environmental variables and dates from IMOS Thredds server (http://thredds.aodn.org.au/)
}
\keyword{internal}
