% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var.R
\name{.extract_var}
\alias{.extract_var}
\title{extract_var}
\usage{
.extract_var(
  unique_positions,
  env_stack,
  env_var,
  .fill_gaps,
  .buffer = NULL,
  verbose = TRUE,
  .full_timeperiod = FALSE
)
}
\arguments{
\item{unique_positions}{dataframe with unique coordinates and dates from where to extract daily environmental data}

\item{env_stack}{a raster stack object with daily spatial environmental layers associated with detection dates (output from \code{.pull_var} function)}

\item{env_var}{name of extracted variable (brought in from \code{extractEnv} function)}

\item{.fill_gaps}{Fill spatial gaps within each layer using a distance buffer}

\item{.buffer}{radius of buffer (in m) around each detection from which environmental variables should be extracted from. A median value of pixels
that fall within the buffer will be used if \code{.fill_gaps = TRUE}}

\item{verbose}{should function provide details of what operation is being conducted. Uses parameter from parent function \code{extractEnv}}
}
\value{
a dataframe with unique position, unique dates and extracted environmental variables
}
\description{
Extract variables from pulled variables
}
\details{
Internal function to conduct raster extraction of environmental variables from downloaded raster stack (\code{env_stack}) and detection
positions (\code{unique_positions})
}
\keyword{internal}
