% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_find.R
\name{ext_find}
\alias{ext_find}
\title{Identify spatial and temporal extents of data}
\usage{
ext_find(.df, .X, .Y, .datetime, .full_timeperiod, verbose)
}
\arguments{
\item{.df}{Dataframe given as input to extract environmental data for}

\item{.X}{Column containing logitudes}

\item{.Y}{Column containing latitudes}

\item{.datetime}{Column containing timestamps}

\item{.full_timeperiod}{Whether data extraction should be performed for the entire period or not}
}
\value{
List of objects
}
\description{
Determines spatial and temporal ranges for data extraction
}
\details{
Internal function to determine spatial and temporal ranges for data extraction
}
\keyword{internal}
