% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_func.R
\name{extract_func}
\alias{extract_func}
\title{Extract IMOS environmental data remotely}
\usage{
extract_func(
  dataset,
  data_details,
  lon.min,
  lon.max,
  lat.min,
  lat.max,
  var_name,
  .fill_gaps,
  .buffer
)
}
\arguments{
\item{dataset}{Environmental data details (URL and netCDF variable names)}

\item{data_details}{Spatial and temporal details of data extraction, created with ext_find}

\item{lon.min}{Minimum longitude for data extraction}

\item{lon.max}{Maximum longitude for data extraction}

\item{lat.min}{Minimum latitude for data extraction}

\item{lat.max}{Maximum latitude for data extraction}

\item{var_name}{Name of environmental variable of interest}

\item{.fill_gaps}{Should NAs be filled?}

\item{.buffer}{Size of the buffer area to be user to fill NAs}
}
\description{
Remotely open and crop IMOS netCDF files and extract data
}
\details{
Internal function to open and extract IMOS environmental data
}
\keyword{internal}
