% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMoor.R
\name{extractMoor}
\alias{extractMoor}
\title{Link the QC Detection Dataset to the Sensor Value obtained by the Nearest IMOS Mooring}
\usage{
extractMoor(
  trackingData,
  file_loc,
  sensorType = "temperature",
  timeMaxh = Inf,
  distMaxkm = Inf,
  targetDepthm = NA,
  scalc = c("min", "max", "mean")
)
}
\arguments{
\item{trackingData}{dataframe containing acoustic detection data in IMOS QC format with the \code{moor_site_code} next to it as generated using the \code{mooringDistance} function.}

\item{file_loc}{character string describing folder location in which NetCDF files of mooring data have been saved after running mooringDownload function}

\item{sensorType}{character string containing name of mooring sensor to query. Can be "temperature", "velocity" or "salinity".}

\item{timeMaxh}{optional numeric string containing the maximum time threshold to merge detection and mooring sensor values.}

\item{distMaxkm}{optional numeric string containing the maximum distance threshold in kilometers to include in output.}

\item{targetDepthm}{extracts the nearest sensor to this depth value. if set as NA then all sensors returned for this timestamp at this mooring}

\item{scalc}{select the lower or higher depth value when there are 2 options for sensors nearest the targetDepth provided. Users can also specify a mean or \code{NA} to return all sensor values.}
}
\value{
the \code{trackingData} dataframe as a nested tibble object with the sensor values from the nearest \code{moorLocations} that fall within the specified time, distance and depth thresholds.
}
\description{
Extracts specified sensor value(s) observed by the IMOS mooring nearest to the supplied QC'd detection location(s)
}
