% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataQC.R
\name{get_data}
\alias{get_data}
\title{read acoustic detection data and metadata from IMOS WebApp-downloaded files}
\usage{
get_data(det = NULL, rmeta = NULL, tmeta = NULL, meas = NULL, logfile)
}
\arguments{
\item{det}{path to detections text file}

\item{rmeta}{path to receiver metadata text file}

\item{tmeta}{path to tag deployment metadata text file}

\item{meas}{path to animal measurements text file}

\item{logfile}{path to logfile; default is the working directory}
}
\value{
a list of data.frames by individual tag deployments ready for QC
}
\description{
accesses files downloaded from the IMOS ATF WebApp and convert them to
a format expected by the IMOS ATF Quality Control functions. For internal use only
}
\examples{

}
\keyword{internal}
