% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataQC_arbitrary.R
\name{get_data_arbitrary}
\alias{get_data_arbitrary}
\title{read acoustic detection data and metadata from an arbitrary data source}
\usage{
get_data_arbitrary(
  det = NULL,
  rmeta = NULL,
  tmeta = NULL,
  meas = NULL,
  logfile,
  checks = c("all"),
  det_id_column = "transmitter_deployment_id",
  tag_id_column = "transmitter_deployment_id",
  det_rcvr_column = "receiver_deployment_id",
  rcvr_id_column = "receiver_deployment_id",
  data_format = "imos",
  col_spec = NULL
)
}
\arguments{
\item{det}{path to detections text file}

\item{rmeta}{path to receiver metadata text file}

\item{tmeta}{path to tag deployment metadata text file}

\item{meas}{path to animal measurements text file}

\item{logfile}{path to logfile; default is the working directory}

\item{checks}{vector containing all of the specific checks you want to turn
on or off while pulling in the data}

\item{det_id_column}{specifies the tag ID variable used for ID checks when tags
are in detections data but not the tag metadata}

\item{tag_id_column}{specifies the tag ID variable used for ID checks when tags
are in detections data but not the tag metadata}

\item{det_rcvr_column}{specifies the receiver ID variable used for ID checks
when receivers are in detections data but not the receiver metadata}

\item{data_format}{the type of format and therefore the type of columns we'll
need to use when we join all this together.}
}
\value{
a list of data.frames by individual tag deployments ready for QC
}
\description{
accesses files from an arbitrary data source using OTN-formatted
data and converts them to the format expected by the remora QC functions.
For internal use only
}
\keyword{<institution}
\keyword{appropriate}
\keyword{be}
\keyword{choosing}
\keyword{columns}
\keyword{data'}
\keyword{file}
\keyword{folder,}
\keyword{for}
\keyword{format}
\keyword{four}
\keyword{in}
\keyword{in.'}
\keyword{internal}
\keyword{is}
\keyword{name>_detections/receiver/transmitter/animal_measurement_columns,}
\keyword{of}
\keyword{referenced}
\keyword{the}
\keyword{these}
\keyword{type}
\keyword{whichever}
\keyword{will}
