% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erddap_ncdf.R
\name{get_erddap_ncdf}
\alias{get_erddap_ncdf}
\title{get environmental netCDF files via an ERDDAP server}
\usage{
get_erddap_ncdf(
  url,
  file,
  layer,
  df,
  date_col = "datecollected",
  lat_col = "latitude",
  lon_col = "longitude"
)
}
\arguments{
\item{url}{...}

\item{file}{...}

\item{layer}{...}

\item{df}{...}

\item{date_col}{name of date column as a character string}

\item{lat_col}{name of latitude column as a character string}

\item{lon_col}{name of longitude column as a character string}
}
\value{
gridded environmental data
}
\description{
For internal use only
}
\keyword{internal}
