% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expert_distribution_shp_qc.R
\name{get_expert_distribution_shp}
\alias{get_expert_distribution_shp}
\title{match species CAAB code with CSIRO expert distribution shapefiles,
when available.}
\usage{
get_expert_distribution_shp(CAAB_species_id, spe)
}
\arguments{
\item{CAAB_species_id}{CAAB id of species for which a distribution
shapefile is required}

\item{spe}{species_scientific_name for which a distribution shapefile
is required}
}
\value{
shp_b is a spatial polygon sf data.frame object of the species' distribution
}
\description{
Get expert distribution shapefile, if available, from CSIRO's
Geoserver. If not, ALA shapefiles (http://www.ala.org.au/) are mathced
against species names and used to validate whether detections occur within
each species' known range.
}
\details{
For a few species acoustically tagged no shapefile exists, for a
few others multiple shapefiles need to be merged into a single one. \code{sf::st_buffer}
extends the boundary of the original shapefile.
}
\keyword{internal}
