% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_shapes.R
\name{get_qc_shapes}
\alias{get_qc_shapes}
\title{Generate the polygon for a species' home range for use in the QC run.}
\usage{
get_qc_shapes(
  scientificName,
  fraction = 1,
  partsCount = 1,
  buff = 5e+05,
  clipToCoast = "aquatic",
  bounds = bounds
)
}
\arguments{
\item{scientificName}{The scientific name of the animal}

\item{fraction}{passed directly through to createPolygon}

\item{partsCount}{passed directly through to createPolygon}

\item{buff}{passed directly through to createPolygon}

\item{clipToCoast}{passed directly through to createPolygon}

\item{bounds}{passed directly through to createPolygon}
}
\value{
The polygon representing the species home range.
}
\description{
Wrap a call to getOccurrence and createPolygon so that the user can supply all their parameters at once and get the polygon back in one call.
}
