% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBuffer.R
\name{getBuffer}
\alias{getBuffer}
\title{Create Buffer to Crop Shapefile}
\usage{
getBuffer(detection_extract, buffer_percentage = 0.15)
}
\arguments{
\item{detection_extract}{A detection extract file containing a latitude/longitude column.}

\item{buffer_percentage}{The percentage (between 0 and 1) to add to the lat/lon of the bounding box, in case the user wants more space around their area.}
}
\value{
returns a buffer as a 4-element vector
}
\description{
Create a buffer from detection data to crop species range shapefile
}
\details{
Creates a buffer for shapefile cropping from location info in the detection_extract
}
