% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistance.R
\name{getDistance}
\alias{getDistance}
\title{Find closest mooring}
\usage{
getDistance(trackingData, moorLocations)
}
\arguments{
\item{trackingData}{dataframe containing acoustic detection data in IMOS QC format}

\item{moorLocations}{dataframe containing the locations of IMOS moorings}
}
\value{
The \code{trackingData} dataframe with the nearest  \code{moor_site_code} to each acoustic detection
}
\description{
Links the QC Detection Dataset to the Nearest IMOS Mooring ID
}
