% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataUrl.R
\name{getMetadataUrl}
\alias{getMetadataUrl}
\title{getMetadataUrl}
\usage{
getMetadataUrl(url, type = "fgdc")
}
\arguments{
\item{url}{...}

\item{type}{...}
}
\value{
a valid URL for metadata
}
\description{
Helper function to grab requisite metadata, given a URL
}
\details{
Internal function to process metadata URL
}
\keyword{internal}
