% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOccurrence.R
\name{getOccurrence}
\alias{getOccurrence}
\title{Get Occurrence File from OBIS and GBIF}
\usage{
getOccurrence(sciName)
}
\arguments{
\item{sciName}{The scientific name of the species you want to query.}
}
\value{
Returns a csv representing occurrence data for the fish in question.
}
\description{
Using a species' scientific name, get a CSV representing occurrence data from the combined datasets of OBIS and GBIF
}
\details{
Takes the scientific name and uses it to get an AphiaID that can then be used to query OBIS and GBIF.
}
