% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOccurrencesFromList.R
\name{getOccurrencesFromList}
\alias{getOccurrencesFromList}
\title{Get A Series of Occurrence Files from OBIS and GBIF}
\usage{
getOccurrencesFromList(
  sciNames,
  save = FALSE,
  saveFolder = "localOccurrenceData/"
)
}
\arguments{
\item{sciNames}{The list of scientific names you want to query (if you only supply one name, it will pass it straight through to getOccurrence)}

\item{save}{Whether or not to save the occurrence files as CSVs locally, for access later.}

\item{saveFolder}{The local location in which to store occurrence data as CSVs.}
}
\value{
Returns a list of dataframes representing occurrence data for the fish in question.
}
\description{
Using a list of species' scientific names, return a list of dataframes containing the occurrence data of the species. This is mostly a wrapper for getOccurrence, I thought it would be more elegant to offload the 'multiple' part
to a different function.
}
\details{
Takes the scientific name and uses it to get an AphiaID that can then be used to query OBIS and GBIF.
}
