% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlap.R
\name{getOverlap}
\alias{getOverlap}
\title{Extract number of tag detections that fall within the duration of sensor coverage at each mooring}
\usage{
getOverlap(x)
}
\arguments{
\item{x}{The \code{trackingData} dataframe with the nearest  \code{moor_site_code} to each acoustic detection}
}
\value{
a grouped tibble describing proportion of temporal overlap between mooring coverage and detections dataset
}
\description{
Extract number of tag detections that fall within the duration of sensor coverage at each mooring
}
