% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPolygonsFromDetectionExtract.R
\name{getPolygonsFromDetectionExtract}
\alias{getPolygonsFromDetectionExtract}
\title{From an OTN-style detection extract, compile the scientific names and load/create/cache appropriate polygons representing home ranges.}
\usage{
getPolygonsFromDetectionExtract(
  extract_path,
  scientificNameColumn = "scientificName"
)
}
\arguments{
\item{extract_path}{A path to the detection extract.}

\item{scientificNameColumn}{If for whatever reason the column your scientific names are stored in isn't 'scientificName',
you can specify a different one.}
}
\value{
A list of polygons that can be passed on to runQC.
}
\description{
This function wraps other polygon generation code. Given a detection extract (and, if necessary, appropriate nonstandard column name,
) it will iterate over the scientific names and create polygons as it goes- loading them if they're already cached, caching them if they're not.
}
