% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesRangePolygon.R
\name{getSpeciesRangePolygon}
\alias{getSpeciesRangePolygon}
\title{title Get Species Range Polygon}
\usage{
getSpeciesRangePolygon(
  occurrence_path,
  occurrence_nas = c("", "NA", "{}"),
  fraction = 0.7,
  buff = 1000,
  partCount = 20,
  coordHeaders = c("decimallongitude", "decimallatitude"),
  clipToCoast = "aquatic"
)
}
\arguments{
\item{occurrence_path}{The path to the occurrence file}

\item{occurrence_nas}{Optional parameter to properly replace NAs in the occurrence file}

\item{fraction}{Pass-through to getDynamicAlphaHull, see that doc for info}

\item{buff}{Same as fraction}

\item{partCount}{Same as fraction}

\item{coordHeaders}{same as Fraction}

\item{clipToCoast}{same as fraction}
}
\value{
The output from getDynamicAlphaHull, which contains a polygon and alpha channel.
}
\description{
Takes an occurrence file (usually from OBIS) representing a fish's home range, and converts it into
an alpha hull that can function as a polygon for QC tests in Remora that require it.
}
