% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabQC.R
\name{grabQC}
\alias{grabQC}
\title{grab subsets of the QC output}
\usage{
grabQC(
  x,
  what = c("dQC", "detections", "QCflags", "tag_meta", "rec_meta", "meas"),
  flag = "all"
)
}
\arguments{
\item{x}{a QC output object (a nested tibble with class \code{remora_QC})}

\item{what}{defined subset of the QC output is to be grabbed; either
\code{detections}, \code{QCflags}, \code{dQC} (detections and QCflags),
\code{tag_meta} (transmitter deployment metadata),
\code{rec_meta} (receiver deployment metadata), or
\code{meas} (animal measurements)}

\item{flag}{specifies which quality controlled detections to return
(see examples): any combination of: \code{valid}, \verb{likely valid},
\verb{likely invalid}, \code{invalid}, or \code{all}. The default is to return all
detections. Ignored if \code{what} is any of \code{tag_meta}, \code{rec_meta} or \code{meas}.}
}
\value{
a data frame with the requested subset of the QC output
}
\description{
\code{grabQC()} lets you obtain subsets of the QC output
}
\examples{
## grab detections and QCflags from example QC output & return only the
## `valid` and `likely valid` detections
data(TownsvilleReefQC)
d.qc <- grabQC(TownsvilleReefQC, what = "dQC", flag = c("valid", "likely valid"))

## return all detections
d.qc <- grabQC(TownsvilleReefQC, what = "dQC")

}
