% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imos_to_otn_receivers.R
\name{imos_to_otn_receivers}
\alias{imos_to_otn_receivers}
\title{Map IMOS receiver metadata to an OTN-like format}
\usage{
imos_to_otn_receivers(rcvr_dataframe)
}
\arguments{
\item{rcvr_dataframe}{A dataframe containing IMOS receiver metadata.}
}
\value{
A dataframe containing the above data in an OTN-like format.
}
\description{
In the same way that otn_imos_column_map takes OTN data and massages it into an IMOS-like format for REMORA,
this function and its ilk take IMOS data (in this case, receiver metadata) and massage it into an OTN-like format, for the
purposes of reporting and more general applicability within the OTN suite of programs.
}
