% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imos_to_otn_tags.R
\name{imos_to_otn_tags}
\alias{imos_to_otn_tags}
\title{Map IMOS tag metadata to an OTN-like format}
\usage{
imos_to_otn_tags(tag_dataframe, animal_measurements_dataframe)
}
\arguments{
\item{tag_dataframe}{A dataframe containing IMOS-formatted tag metadata.}

\item{animal_measurements_dataframe}{A dataframe containing IMOS-formatted animal measurements data.}
}
\value{
A single dataframe containing the tag and measurement data combined into an OTN-like format.
}
\description{
In the same way that otn_imos_column_map takes OTN data and massages it into an IMOS-like format for REMORA,
this function and its ilk take IMOS data (in this case, tag metadata and animal measurements data) and massage it into an OTN-like format, for the
purposes of reporting and more general applicability within the OTN suite of programs.
}
