% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imos_variables.R
\name{imos_variables}
\alias{imos_variables}
\title{Table of all available IMOS variables accessible from \code{remora} package}
\usage{
imos_variables(variable = NULL)
}
\arguments{
\item{variable}{(optional) name of a specific variable you are interested in. Default of \code{NULL} provides details for all available datasets.}
}
\value{
a formatted table with details of available variables accessible through the \code{remora} package
}
\description{
Summary of available IMOS environmental variable to append to detection data.
}
\details{
This function helps users identify what quality controlled environmental variables
can be accessed and appended to their detection data using the \code{\link{extractEnv}} of \code{\link{mooringTable}} functions.
}
\examples{
## Identify all available variables
imos_variables()

## If there is a specific variable you are interested in
imos_variables(variable = "rs_sst_interpolated")

}
