% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs_match.R
\name{locs_match}
\alias{locs_match}
\title{Match extracted IMOS environmental data with detections}
\usage{
locs_match(
  aux_df,
  aux_env,
  X,
  Y,
  datetime,
  full_timeperiod,
  station_name,
  var_name
)
}
\arguments{
\item{aux_df}{Dataset containing detections}

\item{aux_env}{Dateset contatining environmental data}

\item{X}{name of column with X coordinate or longitude (EPSG 4326)}

\item{Y}{name of column with Y coordinate or latitude (EPSG 4326)}

\item{datetime}{name of column with date time stamp (Coordinated Universal}

\item{full_timeperiod}{should environmental variables be extracted for all days}

\item{station_name}{if full_timeperiod = TRUE, please provide the column that identifies the name of the
acoustic stations}

\item{var_name}{Name of the variable of interest}
}
\value{
Dataframe containing extracted environmental data
}
\description{
Help funtion to match environmental and detection datasets. Different workflows if data is extracted for full timeperiods.
}
\details{
Internal help function to match datasets
}
\keyword{internal}
