% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.moor.R
\name{merge.moor}
\alias{merge.moor}
\title{Link the QC Detection Dataset to the Sensor Value obtained by the Nearest IMOS Mooring}
\usage{
\method{merge}{moor}(trackingData, moorData, timeMaxh = Inf, distMaxkm = Inf)
}
\arguments{
\item{trackingData}{dataframe containing acoustic detection data in IMOS QC format. This could also have the moorings site_code next to it as generated using the \code{getDistance} function.}

\item{moorData}{optional dataframe containing the sensor data from a single IMOS mooring.}

\item{timeMaxh}{optional numeric string containing the maximum time threshold in hours to merge detection and mooring sensor values.}

\item{distMaxkm}{optional numeric string containing the maximum distance threshold in km to merge detection and mooring sensor values.}
}
\value{
The \code{trackingData} dataframe with the sensor values from the nearest  \code{moor_site_code}
}
\description{
helper function, not called by user
}
\keyword{internal}
