% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mooringTable.R
\name{mooringTable}
\alias{mooringTable}
\title{Pull the Latest IMOS Moorings Dataset from the Server}
\usage{
mooringTable(sensorType = "temperature")
}
\arguments{
\item{sensorType}{Character string containing name of mooring sensor to query. Can be "temperature" or "velocity"}
}
\value{
dataframe containing the locations of IMOS moorings as csv tibble
}
\description{
Uses a Web link to send a Web Feature Service (WFS) query directly to AODN geoserver
}
