% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{n_detections}
\alias{n_detections}
\title{Number of total detections per transmitter}
\usage{
n_detections(data)
}
\arguments{
\item{data}{detections data}
}
\description{
not to be called by user
}
