% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{n_receivers}
\alias{n_receivers}
\title{Number of total receivers}
\usage{
n_receivers(data)
}
\arguments{
\item{data}{merged detections data and receiver metadata}
}
\description{
not to be called by user
}
