% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDT.r
\name{plotDT}
\alias{plotDT}
\title{Depth-time plot}
\usage{
plotDT(
  moorData,
  moorName,
  dateStart = NULL,
  dateEnd = NULL,
  varName = c("temperature", "vcur", "ucur", "psal"),
  trackingData = NULL,
  speciesID,
  IDtype = "CAAB",
  detStart = NULL,
  detEnd = NULL
)
}
\arguments{
\item{moorData}{Dataframe containing the sensor data from a single IMOS mooring.}

\item{moorName}{Character string specifying mooring name using IMOS ID code.}

\item{dateStart}{Optional character string of start date of date range of mooring data to plot, as "YYYY-MM-DD"}

\item{dateEnd}{Optional character string of end date of date range of mooring data to plot, as "YYYY-MM-DD"}

\item{varName}{Character string of mooring sensor variable of interest. Can be "temperature", "ucur" for meridional (u) velocity, "vcur" for zonal (v) velocity}

\item{trackingData}{Dataframe containing tag detections data to plot. Takes output from getDistance() function as input. Set as NULL for a base plot with no detections.}

\item{speciesID}{Character string describing species or tag identifier.}

\item{IDtype}{Character string describing type of species or tag identifier. Can be CAAB code, tag ID, common name or scientific name. Must match identifier in detections dataset.}

\item{detStart}{Optional character string of start date of date range of detections data to overlay on depth-time plot.}

\item{detEnd}{Optional character string of end date of date range of detections data to overlay on depth-time plot.}
}
\description{
Creates an interactive depth-time plot from a sensor dataset from a specified IMOS mooring, with an option to overlay species detection records from tracking dataset from closest receiver
}
