% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{produce maps showing the occurrence of QC'd detections per species}
\usage{
plotQC(x, path = getwd())
}
\arguments{
\item{x}{a remora output object with \code{class(remora_QC)}}

\item{path}{string; path where maps will be saved. Default is to write plot
to the working directory as a \code{.png} file.}
}
\value{
produces maps showing species expert distribution and
location of QC'd detections
}
\description{
\code{plotQC()} QC'd detections colour-coded by their
assessed validity status, overlaid on species expert distribution extent
}
\examples{
## example QC'd data
data(TownsvilleReefQC)

## Plot QC output to graphics window
plotQC(TownsvilleReefQC, path = NULL)

\dontrun{
## Plot QC output, saved to the working directory as .png file(s) 
##   using species common name(s)
plotQC(TownsvilleReefQC)
}

}
