% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_leaf.R
\name{plotQCint}
\alias{plotQCint}
\title{produce interactive leaflet maps showing the occurrence of QC'd detections per species}
\usage{
plotQCint(x)
}
\arguments{
\item{x}{a remora output object with \code{class(remora_QC)}}
}
\value{
produces interactive leaflet maps of species expert distribution and
location of QC'd detections
}
\description{
\code{plotQCint()} QC'd detections colour-coded by their
assessed validity status, overlaid on species expert distribution extent
}
\examples{
## example QC'd data
data(TownsvilleReefQC)

## render output to viewer
plotQCint(TownsvilleReefQC)


}
