% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_detection_qc.R
\name{qc_detection_qc}
\alias{qc_detection_qc}
\title{QC Detection QC test}
\usage{
qc_detection_qc(qc_result)
}
\arguments{
\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}
}
\value{
adds final Detection QC test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
