% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_remove_nas.R
\name{qc_remove_nas}
\alias{qc_remove_nas}
\title{QC Remove NAs}
\usage{
qc_remove_nas(data)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}
}
\value{
returns the dataframe with the NA values removed.
}
\description{
Small function to remove NA values from a dataframe before the rest of the checks are run on it.
}
\details{
...
}
\keyword{internal}
