% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_det_distro.R
\name{qc_test_det_distro}
\alias{qc_test_det_distro}
\title{QC Detection Distribution test}
\usage{
qc_test_det_distro(data, latlons, qc_results, species_range)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{latlons}{longlat coords of detections}

\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}

\item{species_range}{the expert species distribution shapefile (if avaiable)}
}
\value{
adds Detection Distribution test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
