% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_release_time_diff.R
\name{qc_test_release_time_diff}
\alias{qc_test_release_time_diff}
\title{QC Release Date test}
\usage{
qc_test_release_time_diff(data, qc_results)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}
}
\value{
adds Release Date test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
