% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_velocity.R
\name{qc_test_velocity}
\alias{qc_test_velocity}
\title{QC velocity test}
\usage{
qc_test_velocity(data, qc_results, dist, ...)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}

\item{dist}{output from \code{remora::shortest_dist}}

\item{velocity_threshold}{...}
}
\value{
adds Velocity test outcomes to temporal_outcome object
}
\description{
QC test for unrealistic velocity between acoustic detections.
Requires at least two detections present in detections data.
}
\details{
...
}
\keyword{internal}
