% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_updated.R
\name{qc_updated}
\alias{qc_updated}
\title{test the validity of individual detections for tags detected more than once}
\usage{
qc_updated(
  x,
  Lcheck = TRUE,
  datecolumn = "datecollected",
  speciescolumn = "scientificname",
  installationcolumn = "station",
  logfile
)
}
\arguments{
\item{x}{a list of un-QC'd data for each tag deployment}

\item{Lcheck}{(logical; default TRUE) test for receiver_deployment_latitudes
in N hemisphere at correct to S hemisphere. Set to FALSE for QC on N hemisphere data}

\item{datecolumn}{The name of the column in which the date information is held.}

\item{speciescolumn}{The name of the column in which the species scientific name is held.}

\item{installationcolumn}{The name of the column in which the receiver installation
name is held.}

\item{logfile}{path to logfile; default is the working directory}
}
\value{
temporal_outcome is a list with each element corresponding to a QC'd tag detection file
}
\description{
Subjects tag detections to 8 quality control tests and appends
results to each detection record
}
\details{
...
}
\keyword{internal}
