% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_open.R
\name{remote_open}
\alias{remote_open}
\title{Initiates IMOS variable extraction process}
\usage{
remote_open(
  input,
  data_details,
  env_buffer,
  .fill_gaps,
  .buffer,
  .parallel,
  .ncores,
  var_name
)
}
\arguments{
\item{input}{IMOS netCDF detail dataframe created with remote_urls}

\item{data_details}{Spatial and temporal details of data extraction, created with ext_find}

\item{env_buffer}{distance (in decimal degrees) to expand the study area beyond the coordinates to extract environmental data. Default value is 1°.}

\item{.fill_gaps}{Should NAs be filled?}

\item{.buffer}{Size of the buffer area to be user to fill NAs}

\item{.parallel}{Should data extraction be performed in parallel?}
}
\value{
Dataframe containing environmental data
}
\description{
Help funtion to process data either in parallel or not
}
\details{
Internal help function to initiate IMOS data extraction
}
\keyword{internal}
