% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_urls.R
\name{remote_urls}
\alias{remote_urls}
\title{Create URLs for IMOS netCDF files}
\usage{
remote_urls(input, var_name, .nrt = TRUE, verbose = TRUE)
}
\arguments{
\item{input}{Spatial and temporal details of data, generated using ext_find}

\item{var_name}{Environmental variable of interest}

\item{.nrt}{Should near-real-time be used for current data? Default is TRUE}

\item{verbose}{Defaullt is TRUE}
}
\value{
Dataframe of netCDF URLs
}
\description{
Creates URLs for openning IMOS netCDF files via OPENDAP
}
\details{
Internal function to generate URLs for tidync processing
}
\keyword{internal}
