% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteNCDF.R
\name{remoteNCDF}
\alias{remoteNCDF}
\title{Remotely open BRAN netCDF files}
\usage{
remoteNCDF(year, month, var, depth, lon.min, lon.max, lat.min, lat.max)
}
\arguments{
\item{year}{Year of interest for data download}

\item{month}{Month of interest for data download}

\item{var}{Variable of interest for data download}

\item{depth}{Depth of interest for data download}

\item{lon.min}{Minimum longitude for data download}

\item{lon.max}{Maximum longitude for data download}

\item{lat.min}{Minimum latitude for data download}

\item{lat.max}{Maximum latitude for data download}
}
\description{
Open netCDF files from BRAN and converts them into data frame format.
}
\details{
Returns a dataframe with environmental data
}
