% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReportFunctions.R
\name{sei}
\alias{sei}
\title{Station Efficiency Index}
\usage{
sei(data, date1, date2)
}
\arguments{
\item{data}{merged detections data and receiver metadata}

\item{date1}{start date for SEI calculation}

\item{date2}{end date for SEI calculation}
}
\value{
a tibble
}
\description{
not to be called by user
}
