% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_dist2_qc.R
\name{shortest_dist2}
\alias{shortest_dist2}
\title{find least-cost distance between two locations}
\usage{
shortest_dist2(position, inst, raster = NULL, tr)
}
\arguments{
\item{position}{locations of deployment and subsequent detections as a
data.frame of longitude, latitude}

\item{inst}{installation names of detection (receiver) locations}

\item{raster}{raster of telemetry region coastline, if NULL then the default
\code{Aust_sr} SpatRaster is used}

\item{tr}{transition matrix for calculating least-cost distances (ie. produced
via transition, with geoCorrection applied)}
}
\value{
a 1-column matrix of shortest distances
}
\description{
calculates least-cost distance between two receiver locations
if a land mass lies between them. Only applied on tags detected more than once
}
\details{
...
}
\keyword{internal}
