% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_dist_sf.R
\name{shortest_dist_sf}
\alias{shortest_dist_sf}
\title{find least-cost distance between two locations}
\usage{
shortest_dist_sf(position, inst, raster, tr)
}
\arguments{
\item{position}{locations of deployment and subsequent detections as a
data.frame of longitude, latitude}

\item{inst}{installation names of detection (receiver) locations}

\item{raster}{a \code{terra} SpatRaster of telemetry region coastline}

\item{tr}{transition matrix for calculating least-cost distances (ie. produced
via transition, with geoCorrection applied)}
}
\value{
a 1-column matrix of shortest distances
}
\description{
calculates least-cost distance between two receiver locations
if a land mass lies between them. Only applied on tags detected more than once
}
\details{
...
}
\keyword{internal}
