% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remora-package.R
\docType{data}
\name{TownsvilleReefQC}
\alias{TownsvilleReefQC}
\title{Quality-controlled bull shark detections (5 individuals,
subsampled for efficiency)}
\format{
.RData
}
\description{
Example bull shark acoustic tracking data. Data were sourced from
the Integrated Marine Observing System (IMOS) - IMOS is supported by the
Australian Government through the National Collaborative Research Infrastructure
Strategy and the Super Science Initiative.
}
\keyword{data}
