% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windDir.R
\name{windDir}
\alias{windDir}
\title{Calculate wind direction}
\usage{
windDir(u, v)
}
\arguments{
\item{u}{horizontal (u) wind speed}

\item{v}{vertical (v) wind speed}
}
\value{
Wind direction in degrees clockwise
}
\description{
helper function to calculate wind direction
}
