% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windSpd.R
\name{windSpd}
\alias{windSpd}
\title{Calculate wind speed}
\usage{
windSpd(u, v)
}
\arguments{
\item{u}{horizontal (u) wind speed}

\item{v}{vertical (v) wind speed}
}
\value{
Wind speed in meters per second
}
\description{
helper function to calculate wind speed
}
\keyword{internal}
