% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_qc.R
\name{write_summary}
\alias{write_summary}
\title{extract summary statistics by tag & write to text file}
\usage{
write_summary(x, path = NULL)
}
\arguments{
\item{x}{a nested tibble with class \code{remora_QC} generated by \code{tag_qc})}

\item{path}{path to write QCsummary.csv file}
}
\value{
summary outcome text file is written to user-specified path
}
\description{
extracts tag-deployment-specific QC summary statistics and writes all to a single text file
}
\keyword{internal}
