% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeQC.R
\name{writeQC}
\alias{writeQC}
\title{write a QCsummary.csv file and/or QC processed \code{.csv} files for each tag deployment}
\usage{
writeQC(x, path = NULL, summary = TRUE, csv = TRUE)
}
\arguments{
\item{x}{a nested tibble with class \code{remora_QC} generated by \code{runQC}}

\item{path}{path to write QC'd files to. If NULL the files are written to
the working directory}

\item{summary}{whether to generate a summary \code{.csv} file. If TRUE then
\code{QCsummary.csv} is written to the specified \code{path}}

\item{csv}{whether to generate the individual \code{.csv} files. If TRUE (default) then
each file is written to the specified \code{path}}
}
\value{
.csv files and/or a QCsummary.csv file are written to the specified \code{path}
}
\description{
writes a single QCsummary.csv file and/or tag-deployment-specific \code{temporal_outcome}'s
of quality-controlled detections to \code{.csv} files
}
\details{
takes a \code{remora_QC} nested tibble and writes .csv files
corresponding to each row of the nested tibble. A summary of the QC process
is written to \code{QCsummary.csv}. The summary contains the following information for each QC'd tag
(see references for more details):
\itemize{
\item{total_detections}{the number of raw detections}
\item{detections_before_deployment}{the number of detections recorded prior to deployment date in metadata}
\item{invalid_deployment_location}{logical indicating whether deployment longitude,latitude is valid}
\item{detections_outside_species_range}{the number of detections that occurred outside the species expert distribution}
\item{valid_detections}{the number of detections that passed the QC process (QC flags 1 and 2)}
\item{tracking_duration_days}{the number of days between deployment and the last valid detection}
\item{invalid_velocity}{the number of detections associated with implausible travel speeds}
}
}
\examples{
\dontrun{
## example QC'd data
data(TownsvilleReefQC)
## write QC output
writeQC(TownsvilleReefQC, summary = TRUE)
}

}
\references{
Hoenner X et al. (2018) Australia’s continental-scale acoustic tracking database and its automated quality control process. Sci Data 5, 170206 https://doi.org/10.1038/sdata.2017.206
}
