% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPolygon.R
\name{createPolygon}
\alias{createPolygon}
\title{Create Polygon}
\usage{
createPolygon(
  occurrenceFile,
  fraction = 0.7,
  buffer = 1000,
  partCount = 20,
  coordHeaders = c("decimalLongitude", "decimalLatitude"),
  clipToCoast = "aquatic",
  returnWhole = FALSE
)
}
\arguments{
\item{fraction}{...}

\item{buffer}{...}

\item{partCount}{...}

\item{coordHeaders}{...}

\item{clipToCoast}{...}

\item{returnWhole}{...}

\item{occurenceFile}{...}
}
\value{
retruns a polygon data frame
}
\description{
Create a species range polygon for QC tests from species occurence data
}
\details{
Creates species range polygon for QC tests
}
