% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc_shapes.R
\name{get_qc_shapes}
\alias{get_qc_shapes}
\title{generate shapefiles and transition layers}
\usage{
get_qc_shapes(
  detection_extract,
  shapefile,
  worldimage = "./testDataOTN/NE2_50M_SR.tif"
)
}
\arguments{
\item{detection_extract}{...}

\item{shapefile}{...}

\item{worldimage}{...}
}
\value{
a list of cropped shapefile, transition layer, and world raster
}
\description{
generate shapefiles and transition layers for QC based on user-supplied parameters
}
\keyword{internal}
