% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBuffer.R
\name{getBuffer}
\alias{getBuffer}
\title{Create Buffer to Crop Shapefile}
\usage{
getBuffer(detection_extract, otn_test_data, buffer_percentage = 0.15)
}
\arguments{
\item{detection_extract}{...}

\item{otn_test_data}{we want this as a proper input variable, not a test data set}

\item{buffer_percentage}{...}
}
\value{
returns a buffer as a 4-element vector
}
\description{
Create a buffer from detection data to crop species range shapefile
}
\details{
Internal function to create a buffer for shapefile cropping
}
\keyword{internal}
