% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabQC.R
\name{grabQC}
\alias{grabQC}
\title{grab subsets of the QC output}
\usage{
grabQC(
  x,
  what = c("dQC", "detections", "QCflags", "tag_meta", "rec_meta", "meas")
)
}
\arguments{
\item{x}{a QC output object (a nested tibble)}

\item{what}{defined subset of the QC output is to be grabbed; either
\code{detections}, \code{QCflags}, \code{dQC} (detections and QCflags),
\code{tag_meta} (transmitter deployment metadata),
\code{rec_meta} (receiver deployment metadata), or
\code{meas} (animal measurements)}
}
\value{
a data frame with the requested subset of the QC output
}
\description{
\code{grabQC()} lets you obtain subsets of the QC output
}
\examples{
## grab detections and QCflags from example QC output
data(qc)
d.qc <- grabQC(qc, what = "dQC")

}
