% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{test the validity of individual detections for tags detected more than once}
\usage{
qc(x, Lcheck = TRUE, logfile)
}
\arguments{
\item{x}{a list of un-QC'd data for each tag deployment}

\item{Lcheck}{(logical; default TRUE) test for receiver_deployment_latitudes
in N hemisphere at correct to S hemisphere. Set to FALSE for QC on N hemisphere data}

\item{logfile}{path to logfile; default is the working directory}
}
\value{
temporal_outcome is a list with each element corresponding to a QC'd tag detection file
}
\description{
Subjects tag detections to 8 quality control tests and appends results to each detection record
}
\details{
...
}
\keyword{internal}
