% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_false_detection_test.R
\name{qc_false_detection_test}
\alias{qc_false_detection_test}
\title{QC False Detection Test}
\usage{
qc_false_detection_test(
  data,
  qc_result,
  type = "time_diff",
  pincock_threshold = 3600
)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{type}{...}

\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}

\item{pinkco_tckhreshold}{...}
}
\value{
adds False Detection test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
