% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_release_location_test.R
\name{qc_release_location_test}
\alias{qc_release_location_test}
\title{title Release Location test}
\usage{
qc_release_location_test(
  data,
  qc_result,
  species_range,
  distances,
  latlons,
  distance_threshold = 500
)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{qc_result}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}

\item{species_range}{the expert species distribution shapefile (if avaiable)}

\item{distances}{...}

\item{latlons}{...}

\item{distance_threshold}{...}
}
\value{
adds Detection Distribution test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
