% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_distance.R
\name{qc_test_distance}
\alias{qc_test_distance}
\title{QC Distance test}
\usage{
qc_test_distance(data, qc_results, dist)
}
\arguments{
\item{data}{data formatted by either \code{remora::get_data} or \code{remora::get_data_arbitrary}}

\item{qc_results}{the \code{temporal_outcome} object, holding all QC test results,
generated by \code{remora::qc}}

\item{dist}{output from \code{remora::shortest_dist}}

\item{latlons}{longlat coords of detections}
}
\value{
adds Distance test outcomes to temporal_outcome object
}
\description{
...
}
\details{
...
}
\keyword{internal}
