% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remora-package.R
\docType{package}
\name{remora-package}
\alias{remora-package}
\alias{remora}
\title{\pkg{remora}}
\description{
Rapid Extraction of Marine Observations for Roving Animals
}
\references{
Hoenner, X et al. (2018) Australia’s continental-scale acoustic
tracking database and its automated quality control process. Scientific Data
5, 170206. https://doi.org/10.1038/sdata.2017.206
}
\seealso{
runQC extractEnv extractMoor shinyReport
}
\author{
Ross Dwyer, Xavier Hoenner, Charlie Huveneers, Fabrice Jaine, Ian Jonsen, Francisca Maron, Kylie Scales, Vinay Udyawer
}
\keyword{remora}
