% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runQC}
\alias{runQC}
\title{run IMOS-ATF acoustic detections quality control process}
\usage{
runQC(x, .parallel = FALSE, .ncores = detectCores() - 2, .progress = TRUE)
}
\arguments{
\item{x}{a 4-element list of paths to detections, receiver and transmitter
deployment metadata, and animal measurements data files. These data must be
downloaded from the IMOS-ATF Web App (URL), or have exactly the same
structure and variable names as the Web App data.}

\item{.parallel}{logical; run QC tests in parallel across multiple processors
(default is FALSE)}

\item{.ncores}{integer; number of cores to run in parallel. If NULL and
\code{parallel = TRUE} then process will run across all available cores,
otherwise run across user-specified cores}

\item{.progress}{logical; display QC progress (default is TRUE).}
}
\value{
the QC output is returned to the parent frame as a nested tibble
with class \code{remora_QC}
}
\description{
conduct quality control (QC) on IMOS-ATF acoustic detections data
}
\details{
The QC process merges data from the supplied files downloaded via
the IMOS-ATF Web App (URL): \code{IMOS_detections.csv};
\code{IMOS_receiver_deployment_metadata.csv};
\code{IMOS_transmitter_deployment_metadata.csv};
and \code{IMOS_animal_measurements.csv}. Eight quality control tests are
performed on the detections, as outlined in Hoenner et al. (2018), and
QC flags are appended to the merged data for each of these 8 tests.

The QC flags are values ranging between 1 and 4, representing \code{valid},
\verb{likely valid}, \verb{likely invalid}, and \code{invalid} detections, respectively.
The user can then employ these flags, in various combinations, to filter the
merged data (see \code{examples} & \code{vignette}).

Utility functions are provided to subset the merged data in various ways
from the nested tibble and to visualise the QC results (see \code{examples}
& \code{vignette}).

A QC log is written to \code{QC_logfile.txt} in the working directory. The
logfile documents potential data issues discovered during the QC process:
e.g., \code{transmitter_deployment_id}'s present in the detections
file but not in the transmitter metadata file (if supplied);
\code{receiver_deployment_id}'s present in the detections file but not in the
receiver metadata file (if supplied); NA's present in transmitter deployment
locations. Generally, these issues can not be corrected automatically and
require the user to investigate the cause and take appropriate steps to
correct the data.
}
\examples{
## specify files to QC - use supplied example .csv data
files <- list(det = system.file(file.path("test_data","IMOS_detections.csv"),
                    package = "remora"),
              rmeta = system.file(file.path("test_data","IMOS_receiver_deployment_metadata.csv"),
                    package = "remora"),
              tmeta = system.file(file.path("test_data","IMOS_transmitter_deployment_metadata.csv"),
                    package = "remora"),
              meas = system.file(file.path("test_data","IMOS_animal_measurements.csv"),
                    package = "remora"))
qc.out <- runQC(files)
plotQC(qc.out, path = NULL) # plots to default graphics device

## get detections with QC flags
d.qc <- grabQC(qc.out, what = "dQC")

## clean up
system("rm QC_logfile.txt")

}
\references{
Hoenner, X et al (2018) Australia’s continental-scale acoustic
tracking database and its automated quality control process. Scientific Data
5, 170206. https://doi.org/10.1038/sdata.2017.206
}
