% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReport.R
\name{shinyReport}
\alias{shinyReport}
\title{Render Shiny report}
\usage{
shinyReport(type = "transmitter")
}
\arguments{
\item{type}{"transmitter" or "receiver" to produce the corresponding report}
}
\value{
Shiny app with Transmitter or Receiver Project visualisations and statistics
}
\description{
Renders the Shiny App for transmitter and receiver reports.
}
\examples{
shinyReport(type = "transmitter")
}
